Updated SQL Request for PostTable and changed Post attributes

+++ src/main/java/fr/eseoye/eseoye/beans/PostComplete.java
@@ -8,8 +8,8 @@ import java.sql.Date;
 public class PostComplete extends Post{
     private String content;
 
-    public PostComplete(int id, String title, String author, int price, Date date, String content) {
-        super(id, title, author,price, date);
+    public PostComplete(int id, String title, String author, float price, Date date, String content) {
+        super(id, title, author, price, date);
         this.content = content;
     }
